pyproject.toml
[project]
name = "course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "minsearch>=0.0.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydantic-ai>=1.32.0",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
